# backend/internal/store/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package store

import (
	"time"

	"github.com/jackc/pgx/v5/pgtype"
)

type Answer struct {
	ID               string  `json:"id"`
	AttemptID        string  `json:"attemptId"`
	QuestionID       string  `json:"questionId"`
	SelectedOptionID *string `json:"selectedOptionId"`
	Correct          *bool   `json:"correct"`
}

type Credential struct {
	ID           string `json:"id"`
	ParentID     string `json:"parentId"`
	CredentialID []byte `json:"credentialId"`
	PublicKey    []byte `json:"publicKey"`
	Counter      int64  `json:"counter"`
	IsBackup     bool   `json:"isBackup"`
}

type ExamAttempt struct {
	ID               string             `json:"id"`
	ExamDefinitionID string             `json:"examDefinitionId"`
	StudentID        string             `json:"studentId"`
	Status           string             `json:"status"`
	QuestionIds      []string           `json:"questionIds"`
	ScorePct         *int32             `json:"scorePct"`
	Passed           *bool              `json:"passed"`
	PerTopic         []byte             `json:"perTopic"`
	CooldownUntil    pgtype.Timestamptz `json:"cooldownUntil"`
	StartedAt        time.Time          `json:"startedAt"`
	SubmittedAt      pgtype.Timestamptz `json:"submittedAt"`
}

type ExamDefinition struct {
	ID            string    `json:"id"`
	SubjectID     string    `json:"subjectId"`
	Name          string    `json:"name"`
	Type          string    `json:"type"`
	ScopeTopicIds []string  `json:"scopeTopicIds"`
	Size          int32     `json:"size"`
	PassBar       int32     `json:"passBar"`
	CooldownMin   int32     `json:"cooldownMin"`
	RewardStyle   string    `json:"rewardStyle"`
	CreatedAt     time.Time `json:"createdAt"`
}

type Option struct {
	ID         string `json:"id"`
	QuestionID string `json:"questionId"`
	Text       string `json:"text"`
	Order      int32  `json:"order"`
}

type Parent struct {
	ID          string    `json:"id"`
	DisplayName string    `json:"displayName"`
	Email       string    `json:"email"`
	CreatedAt   time.Time `json:"createdAt"`
}

type Question struct {
	ID              string    `json:"id"`
	SubjectID       string    `json:"subjectId"`
	TopicID         *string   `json:"topicId"`
	Text            string    `json:"text"`
	CorrectOptionID string    `json:"correctOptionId"`
	Difficulty      string    `json:"difficulty"`
	Enabled         bool      `json:"enabled"`
	CreatedAt       time.Time `json:"createdAt"`
}

type Setting struct {
	ID                 string `json:"id"`
	RewardRateMinPerQ  int32  `json:"rewardRateMinPerQ"`
	DailyCapHours      int32  `json:"dailyCapHours"`
	DefaultExamSize    int32  `json:"defaultExamSize"`
	DefaultPassBar     int32  `json:"defaultPassBar"`
	DefaultCooldownMin int32  `json:"defaultCooldownMin"`
	KnowledgeBackend   string `json:"knowledgeBackend"`
	DifficultyRamp     bool   `json:"difficultyRamp"`
}

type Source struct {
	ID        string    `json:"id"`
	SubjectID string    `json:"subjectId"`
	Type      string    `json:"type"`
	Title     string    `json:"title"`
	Status    string    `json:"status"`
	FileRef   *string   `json:"fileRef"`
	Url       *string   `json:"url"`
	Text      *string   `json:"text"`
	CreatedAt time.Time `json:"createdAt"`
}

type Student struct {
	ID         string    `json:"id"`
	Name       string    `json:"name"`
	GradeLevel *string   `json:"gradeLevel"`
	AvatarUrl  *string   `json:"avatarUrl"`
	Notes      *string   `json:"notes"`
	PinHash    *string   `json:"pinHash"`
	CreatedAt  time.Time `json:"createdAt"`
}

type Subject struct {
	ID          string    `json:"id"`
	Name        string    `json:"name"`
	Color       *string   `json:"color"`
	Icon        *string   `json:"icon"`
	Description *string   `json:"description"`
	Archived    bool      `json:"archived"`
	CreatedAt   time.Time `json:"createdAt"`
}

type Topic struct {
	ID        string  `json:"id"`
	SubjectID string  `json:"subjectId"`
	Name      string  `json:"name"`
	SourceID  *string `json:"sourceId"`
	PageStart *int32  `json:"pageStart"`
	PageEnd   *int32  `json:"pageEnd"`
	Order     int32   `json:"order"`
	Active    bool    `json:"active"`
}