# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: StudyRover API
  version: 0.1.0
  description: StudyRover Phase 1 — the tutoring core. A parent sets up subjects/topics/questions/exams;
    a student takes exams; the system produces scores + mastery and emits a ScoreEvent. JSON over HTTP;
    parent cookie session required except for the student exam + student-auth endpoints.
  license:
    name: Proprietary
servers:
- url: /api
  description: StudyRover API, served by the Go binary under the /api base path.
security:
- cookieSession: []
...
paths:
...
  /topics/{id}:
    parameters:
    - $ref: '#/components/parameters/IdPath'
    get:
      operationId: getTopic
      tags:
      - topics
      summary: Get a topic
      responses:
        '200':
          description: The topic.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Topic'
        default:
          $ref: '#/components/responses/Problem'
    put:
      operationId: updateTopic
      tags:
      - topics
      summary: Update a topic (partial)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Topic'
      responses:
        '200':
          description: The updated topic.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Topic'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteTopic
      tags:
      - topics
      summary: Delete a topic
      responses:
        '204':
          description: The topic was deleted.
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: sr_session
      description: HTTP-only session cookie issued by /auth/* after a successful WebAuthn (parent) or
        PIN (student) ceremony. Carries the role + subject id.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      description: Resource identifier (UUID/cuid).
      schema:
        type: string
...
  responses:
    Problem:
      description: An error occurred; the body is an RFC 7807-style problem.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Code:
      type: string
      description: 'Machine-readable error category. Maps to the matching HTTP status: VALIDATION->422/400,
        NOT_FOUND->404, UNAUTHORIZED->401, CONFLICT->409, INTERNAL->500.'
      enum:
      - VALIDATION
      - NOT_FOUND
      - UNAUTHORIZED
      - CONFLICT
      - INTERNAL
    Problem:
      type: object
      description: RFC 7807-style problem detail returned for any error response. The HTTP status of the
        response matches the `status` field and the `code` category.
      additionalProperties: false
      required:
      - type
      - title
      - status
      - code
      properties:
        type:
          type: string
          description: A URI reference identifying the problem type. May be the literal "about:blank"
            when no specific type is defined.
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          description: The HTTP status code for this occurrence of the problem.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence.
        code:
          $ref: '#/components/schemas/Code'
...
    Topic:
      type: object
      description: A unit of study within a subject. May optionally point at a source and a page range.
        `order` controls syllabus position; `active` toggles whether the topic is in scope for exams.
      additionalProperties: false
      required:
      - id
      - subjectId
      - name
      - order
      - active
      properties:
        id:
          type: string
          description: Unique topic identifier (UUID/cuid).
        subjectId:
          type: string
          description: Identifier of the owning subject.
        name:
          type: string
          description: Human-readable topic name.
          minLength: 1
        sourceId:
          type: string
          description: Optional identifier of the source this topic is drawn from.
        pageStart:
          type: integer
          format: int32
          description: Optional first page of the topic within its source.
          minimum: 1
        pageEnd:
          type: integer
          format: int32
          description: Optional last page of the topic within its source. If both pages are present, pageEnd
            must be >= pageStart.
          minimum: 1
        order:
          type: integer
          format: int32
          description: Position of the topic in the subject's syllabus.
        active:
          type: boolean
          description: Whether the topic is active (in scope for exams).
          default: true
...